Create open knowledge manifest for syringe pump
I'm working with a group of people to make hardware easy to discover.

Named [Open Know How](https://openknowhow.org/) we've come up with a system where a simple file with metadata about a certain project would live in the project documentation page.

From there we are working to create a web-crawler that is going to find these files and index them (the ultimate goal would be to have something like a "hardware" tab on google search (as other search engines).

I took the liberty of creating one of these files for your project, with the bare minimum for it to be indexed in the future! I hope you don't mind! (it is just this .yml file that should live in the root of the repository)

You can find more technical information about Open Know How [here](https://app.standardsrepo.com/MakerNetAlliance/OpenKnowHow/wiki)